app/requirements.txt
# Core API
fastapi
uvicorn[standard]
requests
python-dotenv

# Database - PostgreSQL
sqlalchemy>=2.0
asyncpg

# Database - MongoDB
pymongo[srv]
dnspython

# Cache
redis

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
cryptography
